fusor server: create hostgroups and activation key for rhev deployment
This commit will create hostgroups for a rhev deployment
based upon the: deployment and content management actions
that are performed for the deployment.

E.g. create hostgroup X with lifecyle environment A,
content view B, puppet environment C, puppet classes M, N, O...etc

It will also create an activation key and associate it
to the hostgroups.  This will enable hosts that are provisioned
with these groups to get the proper subscriptions.