fusor server: enable puppet smart class parameter overrides
This commit provides an update that will enable the 'override' flag
on puppet class smart class parameters, based upon the user's configuration
(fusor.yaml).   An initial set of parameters have been specified with
the commit as well.

Since we have less than 10 parameters specified and the cost of
checking/performing the update is low, it is currently being done as part
of the deploy action.  That said, we might want to look at enabling
during a seed (plugin or installer) in the future.