Always retry if we couldn't make the connection.
There are some errors that are truly recoverable, others are not for example no
route to host. But some of these "non recoverable" errors happen because of
timing issues i.e. dns record not updated, ssh daemon not up yet, etc. So if
we always retry, the hope is we'll be able to connect before we exhaust our
retry counter.